requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
fastapi-users[sqlalchemy]==12.1.2
python-jose[cryptography]==3.3.0

# Logging
loguru==0.7.2

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# HTTP client for external APIs
httpx==0.25.2

# Caching and task queue (for future use)
redis==5.0.1
celery==5.3.4

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0